Creating RPG Stats and Bug Fixes
Changes:
Created obj_cbt_stats that will house all the stats for the character and is persistent between rooms
Created a GUI that displays the stats of the character on screen
Created spr_expr and obj_expr that represent collectable EXP for the character
Rearranged and added more enemies to combat test room
Modified scr_attack_state to use the damage of the character
Added a collision event to obj_cbt to collect obj_expr that modifies the variables within obj_cbt_stats
Fixed a bug with dashing where the character always dashed to the right when standing still
Fixed a bug where the obj_cbt's sprite would not return back to the walking sprite after attacking while standing still
Rearranged Room load order to test GUI

Notes:
obj_cbt_stats is a persistent object and must be present in the first room of the game.
I must modify the code to only display the stats on screen when obj_cbt is present so that we can go back to using the main menu and house obj_cbt_stats there.
From there, we can modify the main menu to rewrite the variables of obj_cbt_stats when we select New Game back to the stats it had at level 1.
Then, make it so selecting Start will load us back into the game with our current stats.
This is a stepping stone towards easily implementing a save game feature.